backend/requirements.txt
fastapi
uvicorn[standard]
motor
beanie
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
supabase
